[@kbn/handlebars] Split index.ts into several files (#150230)
An attempt to make the code a little bit easier to parse

(cherry picked from commit eafa5e7f0536715e26d71c3e104525ddd7551a14)

# Conflicts:
#	src/plugins/vis_types/timeseries/public/application/components/lib/replace_vars.ts
