+++ packages/kbn-handlebars/src/symbols.ts
@@ -1,0 +1,8 @@
+/*
+ * Elasticsearch B.V licenses this file to you under the MIT License.
+ * See `packages/kbn-handlebars/LICENSE` for more information.
+ */
+
+export const kHelper = Symbol('helper');
+export const kAmbiguous = Symbol('ambiguous');
+export const kSimple = Symbol('simple');
